requirements.txt
streamlit
Pillow
boto3
python-dotenv
streamlit-drawable-canvas